feat: adding queue calls in controllers and updated tests accordingly. unified services
